+++ src/shared/schema/types/model.type.mart_join_models.schema.d.ts
@@ -581,6 +581,10 @@ export type SchemaModelSelectCTE =
       include?: [SchemaColumnName, ...SchemaColumnName[]];
     };
 /**
+ * Will prevent the automatic date filters from getting added
+ */
+export type SchemaModelExcludeDateFilter = boolean;
+/**
  * Validate model group by
  */
 export type SchemaModelGroupBy1 =
@@ -879,6 +883,7 @@ export interface SchemaModelCTE {
       | SchemaModelSelectCTE
     )[],
   ];
+  exclude_date_filter?: SchemaModelExcludeDateFilter;
   where?: SchemaModelWhere;
   group_by?: SchemaModelGroupBy1;
   having?: SchemaModelHaving;